+++ react/B2BLastOrders.tsx
@@ -1,2 +1,2 @@
-import B2BLastOrders from './components/B2BLastOrders'
+import B2BLastOrders from './components/B2BLastOrders/B2BLastOrders'
 export default B2BLastOrders