+++ react/B2BRepresentativeArea.tsx
@@ -1,2 +1,2 @@
-import B2BRepresentativeArea from './components/B2BRepresentativeArea'
+import B2BRepresentativeArea from './components/B2BRepresentativeArea/B2BRepresentativeArea'
 export default B2BRepresentativeArea