+++ react/components/SkuContent.tsx
@@ -1,11 +1,12 @@
-import React, { ReactNode } from 'react'
+import classNames from 'classnames'
+import React, { ReactNode, SyntheticEvent } from 'react'
 import { useCssHandles } from 'vtex.css-handles'
+import { useProductDispatch } from 'vtex.product-context/ProductDispatchContext'
+import useProduct from 'vtex.product-context/useProduct'
 import { ExtensionPoint } from 'vtex.render-runtime'
+
 import { Item, Product, Seller } from '../typings'
 import { SkuProvider } from './SkuContext'
-import useProduct from 'vtex.product-context/useProduct'
-import { useProductDispatch } from 'vtex.product-context/ProductDispatchContext'
-import classNames from 'classnames'
 
 const CSS_HANDLES = ['skuContentWrapper', 'selectedSkuContentWrapper'] as const
 
@@ -19,8 +20,12 @@ const SkuContent = ({ item, product, children }: Props) => {
   const handles = useCssHandles(CSS_HANDLES)
   const { selectedItem }: { selectedItem: Item } = useProduct()
   const dispatch = useProductDispatch()
-  const handleClick = () => {
-    if (dispatch) {
+  const handleClick = (e: SyntheticEvent) => {
+    const clickedElement = e.target as HTMLElement
+    const isNumericStepperClicked = clickedElement.classList.value.includes(
+      'vtex-numeric-stepper'
+    )
+    if (dispatch && isNumericStepperClicked) {
       dispatch({ type: 'SET_SELECTED_ITEM', args: { item: item } })
     }
   }
@@ -43,7 +48,8 @@ const SkuContent = ({ item, product, children }: Props) => {
         onClick={handleClick}
         role="button"
         tabIndex={parseInt(item.itemId)}
-        onKeyDown={handleClick}>
+        onKeyDown={handleClick}
+      >
         {children}
         {item.sellers.map((seller: Seller) => (
           <ExtensionPoint